Allow the user to specify a fixed guard to use for a new circuit.
Also, this exposes an independently usable method to sample a new
guard without building a complete circuit. Based on this
functionality, the library can work with a guard set handler
implemented by the user.

Note that, if a fixed guard is specified, the order of selecting
relays changes, as exits have to adhere to the IP requirements
imposed by the given guard.